requirements.txt
pybrain3
numpy
Django==1.11.5
django-cors-headers==2.1.0
django-filter==1.0.4
django-sslserver==0.20
djangorestframework==3.6.4
pytz==2017.3
